requirements.txt
web3==6.20.0
requests==2.31.0
python-telegram-bot==21.3
python-dotenv==1.0.1